Shows the area of the skills before being cast.
Closes #2.
If you press the uppercase letter corresponding to a skill, it will
change the hovered background to let you know if that tile is a valid
target or not for the skill.

Skills with no area, such as Sonic Waves, are shown as invalid. This
might be a bug or a feature, but definitely it's an issue that should
be discussed.